connect user api

+++ sechand_backend/user/views.py
@@ -5,7 +5,7 @@ from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
 from django.utils.encoding import force_str
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated, AllowAny
-from .models import VerifyEmailCode, ResetPasswordCode
+from .models import VerifyEmailCode, ResetPasswordCode, Address
 from .utils import send_verify_email, send_reset_password
 from .serializers import CustomUserSerializer, VerifyEmailCodeSerializer, ResetPasswordCodeSerializer
 
@@ -17,18 +17,32 @@ UserModel = get_user_model()
 def custom_login(request):
     username = request.data.get('username')
     password = request.data.get('password')
-    user = authenticate(username=username, password=password)
-    if user is not None:
-        if user.is_verified:
-            login(request, user)
-            serializer = CustomUserSerializer(user)
-            return JsonResponse(serializer.data, status=201)
+    print(username, password)
+    try:
+        user = UserModel.objects.get(username=username)
+        if user.check_password(password):
+            if user.is_verified:
+                login(request, user)
+                serializer = CustomUserSerializer(user)
+                # registered, correct password
+                return JsonResponse({'registered': True, 'success': True, 'userInfo': serializer.data})
+            else:
+                # User is not active, prompt for email verification
+                user.delete()
+                # registered, not verified
+                return JsonResponse({'registered': True, 'success': False, 'userInfo': None})
         else:
-            # User is not active, prompt for email verification
-            user.delete()
-            return JsonResponse({'message': 'Please verify your email before logging in.'}, status=401)
-    else:
-        return JsonResponse({'message': 'Invalid username or password.'}, status=401)
+            # registered, incorrect password
+            return JsonResponse({'registered': True, 'success': False, 'userInfo': None})
+    except UserModel.DoesNotExist:
+        # 3.2. not registered
+        return JsonResponse({'registered': False, 'success': False, 'userInfo': None})
+    
+    # 1. registered, correct password
+    # 2. registered, incorrect password
+    # 3. not registered
+    #   3.1. registered, not verified
+    #   3.2. not registered
 
 
 @api_view(['POST'])
@@ -37,19 +51,49 @@ def register(request):
     username = request.data.get('username')
     password = request.data.get('password')
     email = request.data.get('email')
+    phone = request.data.get('phone')
+    displayname = request.data.get('displayname')
+    address_id = request.data.get('address')
+    address = Address.objects.get(id=address_id)
+    image = request.data.get('image')
+    is_visible = request.data.get('is_visible')
     print(username, password, email)
-    # Create the user
     try:
-        user = UserModel.objects.create_user(username=username, email=email, password=password)
+        user = UserModel.objects.get(username=username)
+        if not user.is_verified:
+            # Update the user
+            user.username = username
+            user.set_password(password)
+            user.email = email
+            user.phone = phone
+            user.displayname = displayname
+            user.address = address
+            user.image = image
+            user.is_visible = is_visible
+            user.save()
+            # Send verification email with the code
+            try:
+                send_verify_email(user)
+                serializer = CustomUserSerializer(user)
+                # registered, not verified
+                return JsonResponse({"new_user": True, "user": serializer.data})
+            except Exception as e:
+                return JsonResponse({'error': str(e)}, status=500)
+        else:
+            # registered and verified
+            return JsonResponse({"new_user": False, "user": None})
+    except UserModel.DoesNotExist:
+        # Create the user
+        user = UserModel.objects.create_user(username=username, email=email, password=password, phone=phone,
+                                             displayname=displayname, address=address, image=image, is_visible=is_visible)
         # Send verification email with the code
         try:
             send_verify_email(user)
             serializer = CustomUserSerializer(user)
-            return JsonResponse(serializer.data, status=201)
+            # not registered, not verified
+            return JsonResponse({"new_user": True, "user": serializer.data})
         except Exception as e:
             return JsonResponse({'error': str(e)}, status=500)
-    except ValueError as e:
-        return JsonResponse({'error': str(e)}, status=400)
 
 
 @api_view(['POST'])
@@ -58,20 +102,31 @@ def verify_email(request, uid):
     # user = request.user
     code = request.data.get('code')
     try:
-        # uid = force_text(urlsafe_base64_decode(uidb64))
         user = UserModel.objects.get(pk=uid)
-        verification_code = VerifyEmailCode.objects.get(user=user)
-        if verification_code.code != code:
-            return JsonResponse({'message': 'Invalid verification code.'}, status=400)
-        if verification_code.is_expired():
-            return JsonResponse({'message': 'The verification code has expired.'}, status=400)
-        # Update the user's status to verified
-        user.is_verified = True
-        user.save()
-        verification_code.delete()
-        return JsonResponse({'message': 'Email verified successfully.'})
-    except VerifyEmailCode.DoesNotExist:
-        return JsonResponse({'message': 'User does not exist.'}, status=404)
+        if not user.is_verified:
+            try:
+                verification_code = VerifyEmailCode.objects.get(user=user)
+                if verification_code.code != code:
+                    # code is incorrect
+                    return JsonResponse({'expired': False, 'correct': False})
+                if verification_code.is_expired():
+                    # code has expired
+                    return JsonResponse({'expired': True, 'correct': True})
+                # Update the user's status to verified
+                user.is_verified = True
+                user.save()
+                verification_code.delete()
+                # code is verified
+                return JsonResponse({'expired': False, 'correct': True})
+            except VerifyEmailCode.DoesNotExist:
+                # code has never been sent
+                return JsonResponse({'message': 'Invalid verification code.'}, status=404)
+        else:
+            # email has been verified
+            return JsonResponse({'message': 'Email already verified.'}, status=400)
+    except UserModel.DoesNotExist:
+        # user does not exist
+        return JsonResponse({'message': 'Invalid user.'}, status=404)
 
 
 @api_view(['POST'])
@@ -81,53 +136,78 @@ def forgot_password(request):
     try:
         user = UserModel.objects.get(email=email)
         try:
-            send_reset_password(user)
-            serializer = CustomUserSerializer(user)
-            return JsonResponse(serializer.data, status=201)
+            if user.is_verified:
+                send_reset_password(user)
+                # success
+                return JsonResponse({"registered": True, "sent": True, "uid": user.id})
+            else:
+                # user registered, not verified
+                return JsonResponse({"registered": False, "sent": False})
         except Exception as e:
-            return JsonResponse({'error': str(e)}, status=500)
+            return JsonResponse({"registered": False, "sent": False}, status=400)
     except UserModel.DoesNotExist:
-        return JsonResponse({'message': 'User email does not exist.'}, status=404)
+        # emial not registered
+        return JsonResponse({"registered": False, "sent": False})
+
+
+# @api_view(['POST'])
+# @permission_classes([AllowAny])
+# def verify_code(request, uid):
+#     code = request.data.get('code')
+#     try:
+#         code = request.data.get('code')
+#         # uid = force_str(urlsafe_base64_decode(uidb64))
+#         user = UserModel.objects.get(pk=uid)
+#         try:
+#             verification_code = ResetPasswordCode.objects.get(user=user)
+#             if verification_code.code != code:
+#                 return JsonResponse({'message': 'Invalid verification code.'}, status=400)
+#             if verification_code.is_expired():
+#                 return JsonResponse({'message': 'The verification code has expired.'}, status=400)
+#             serializer = ResetPasswordCodeSerializer(verification_code)
+#             return JsonResponse(serializer.data, status=201)
+#         except ResetPasswordCode.DoesNotExist:
+#             return JsonResponse({'message': 'Invalid verification code.'}, status=404)
+#     except ResetPasswordCode.DoesNotExist or UserModel.DoesNotExist:
+#         return JsonResponse({'message': 'Invalid verification code.'}, status=404)
 
 
 @api_view(['POST'])
 @permission_classes([AllowAny])
-def verify_code(request, uid):
+def reset_password(request, jhed):
     code = request.data.get('code')
     try:
-        code = request.data.get('code')
-        # uid = force_str(urlsafe_base64_decode(uidb64))
-        user = UserModel.objects.get(pk=uid)
+        # Update the user's password
+        user = UserModel.objects.get(username=jhed)
         verification_code = ResetPasswordCode.objects.get(user=user)
         if verification_code.code != code:
-            return JsonResponse({'message': 'Invalid verification code.'}, status=400)
+            return JsonResponse({'expired': False, 'correct': False})
         if verification_code.is_expired():
-            return JsonResponse({'message': 'The verification code has expired.'}, status=400)
-        serializer = ResetPasswordCodeSerializer(verification_code)
-        return JsonResponse(serializer.data, status=201)
-    except ResetPasswordCode.DoesNotExist or UserModel.DoesNotExist:
-        return JsonResponse({'message': 'Invalid verification code.'}, status=404)
-
-
-@api_view(['POST'])
-@permission_classes([AllowAny])
-def reset_password(request, uid, token):
-    try:
-        # Update the user's password
-        user = UserModel.objects.get(pk=uid)
-        verification_code = ResetPasswordCode.objects.get(user=user)
-        if verification_code.token != token:
-            return JsonResponse({'message': 'Invalid token.'}, status=400)
+            return JsonResponse({'expired': True, 'correct': True})
         new_password = request.data.get('new_password')
         user.set_password(new_password)
         user.save()
         verification_code.delete()
-        return JsonResponse({'message': 'Password reset successfully.'}, status=201)
+        # password updated
+        return JsonResponse({'expired': False, 'correct': True})
     except UserModel.DoesNotExist:
+        # new user
         return JsonResponse({'message': 'Invalid user.'}, status=404)
 
 
 @api_view(['GET'])
-@permission_classes([IsAuthenticated])
-def protected_view(request):
-    return JsonResponse(data={"message": "Hello, world!"})
+def get_user_profile(request):
+    user = request.user
+    serializer = CustomUserSerializer(user)
+    return JsonResponse(serializer.data, status=200)
+
+
+@api_view(['PATCH'])
+def update_user_profile(request):
+    user = request.user
+    serializer = CustomUserSerializer(user, data=request.data, partial=True)
+    if serializer.is_valid():
+        serializer.save()
+        return JsonResponse(serializer.data, status=200)
+    else:
+        return JsonResponse(serializer.errors, status=400)
